Make webkit_modules a split static library in non-component builds.
Previously this was force to a source set because it was too large for official Windows builds to handle as a static library.

This uses the split_static_library template for non-component official Windows builds so it can be a static library now.

BUG=627637

Review-Url: https://codereview.chromium.org/2145253003
Cr-Original-Commit-Position: refs/heads/master@{#405618}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb0a50c30fdf3d7ca8f978fb55f363cf1f702b73