[Layout API] Fix layoutViewItem() null check in Element.cpp
There are two instances where we are creating a LayoutViewItem and
incorrectly checking whether the underlying layout view is null. At
this stage of the block layout API, the correct way is to call the
isNull() method; checking whether the layoutViewItem object itself
is null does not tell you whether the underlying layout view is null.

BUG=499321

Review-Url: https://codereview.chromium.org/2148043003
Cr-Original-Commit-Position: refs/heads/master@{#405645}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e932c6f79609ca3163b40212835d37e8ff1a779