Push API: Implement and ship PushSubscription.options
Intent to implement and ship:
https://groups.google.com/a/chromium.org/d/topic/blink-dev/ysEXU21Dn1Y/discussion

Recent updates to the Push API renamed the PushSubscriptionOptions
dictionary to PushSubscriptionOptionsInit, and instead introduced a
PushSubscriptionOptions interface which should be exposed as an attribute
on PushSubscription objects, reflecting the options that were passed in
when subscribing.

See commits:
https://github.com/w3c/push-api/commit/3f7aaed45680818d6c472e6b6bce8854e0c9bff2
https://github.com/w3c/push-api/commit/5c25fee2dcaecb140d78566ac26207ee602a6844
https://github.com/w3c/push-api/commit/428f4c5c8993fc6af82f1ea9194e1bb1cbec2feb

This patch implements those changes.

BUG=626627

Review-Url: https://codereview.chromium.org/2133673002
Cr-Original-Commit-Position: refs/heads/master@{#405752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8882a829851f9bfe6e1857c9d8b0a7c679d89c00