When fetching try jobs for a CL, include only the latest for each builder.
This changes webkitpy.common.net.rietveld.latest_try_jobs to return only
the latest try job for each builder on the latest patchset; previously
it would return all try jobs for the latest patchset, even if there are
multiple retries for the same builder.

This is the behavior I believe we want for rebaselining from try jobs,
because if some flaky tests fail in a try job on one try, that builder
can be retried, and then new baselines should be downloaded only for
the latest run for that builder.

BUG=474273

Review-Url: https://codereview.chromium.org/2149453002
Cr-Original-Commit-Position: refs/heads/master@{#405796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8fa41e4ad43d97dff7184069da9d169b3cd1aab