Wait until after layout when restoring scroll on exiting fullscreen.
Many pages can change while in fullscreen. For example, youtube applies
display:none to elements not in the fullscreen ancestor chain using CSS
selectors. This means that until we update style and layout, the page will not
be scrollable to the same extent (or at all) as it was when we entered
fullscreen.

The solution is to take note when we exit fullscreen and try to restore the
scale and scroll once the first layout after exiting fullscreen is completed.
This ensures the page is completely back to it's pre-fullscreen state.

BUG=625683

Review-Url: https://codereview.chromium.org/2122013003
Cr-Original-Commit-Position: refs/heads/master@{#405854}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3fb3a358ce9c85bfd7bcda882a23179ac064775