Use SkCanvas transform instead of flipSkImageVertically
Currently in ImageBitmap, whenever we need to flip a SkImage, we call the
function flipSkImageVertically, which read the pixels of the SkImage,
swap the pixels and construct a new SkImage from the swapped pixels again.
This process is un-necessary in the case where we already have a SkCanvas
and are drawing to that canvas. In this case, we can just call the
equivalent transformation on SkCanvas to achieve the same flipY effect.

Review-Url: https://codereview.chromium.org/2144733011
Cr-Original-Commit-Position: refs/heads/master@{#406102}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 433fb76ec9ee583017b02903c958f0af5a022829