[DOM][Editing][CodeHealth] Make private Node::hasEditableStyle static local
Node has two |hasEditableStyle|-named functions and it is very confusing.
Since one is private and uses only Node-public functions, we should make it
 file local function.
Unfortunately public |hasEditableStyle| uses it in the Node header
so we need move it to  Node.cpp.

This is preparation to move Node::*editable* functions to core/editing

BUG=

Review-Url: https://codereview.chromium.org/2155623002
Cr-Original-Commit-Position: refs/heads/master@{#406412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0793f8f201f0a8468e6e3dc5b1a085e4161a684d