[css-grid] repeat() syntax should take a <track-list> argument
The repeat() notation used to allow just 1 <track-size> as second argument.
It was recently changed to support <track-list> instead, meaning that we can
pass an arbitrary number of track sizes and line numbers.

It has been working for some time for repeat() if the first argument was
a positive integer, but it requires some changes for the auto repeat cases
(auto-fill and auto-fit).

BUG=619930

Review-Url: https://codereview.chromium.org/2161003003
Cr-Original-Commit-Position: refs/heads/master@{#406540}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e06bcde925d051ebdcb42243dc15084d762e17b