[Stream] Notify ReadObserver of Stream abortion
Currently there is no means to get notified when a content::Stream is aborted.
As a result, when a fetch body stream created in a ServiceWorker is get
errored, that error is not notified to the page. This CL fixes that problem by
calling ReadObserver::OnDataAvailable when a content::Stream is aborted.

BUG=627046

Review-Url: https://codereview.chromium.org/2146773002
Cr-Original-Commit-Position: refs/heads/master@{#406553}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ecb30baac7eed72a9729e0f7cdf84fe8cd7178d