EME: Update tests so 'audioCapabilities' always provided
EME spec now requires that one of audioCapabilities or
videoCapabilities be specified when calling
requestMediaKeySystemAccess, so update the tests to always
provide audioCapabilities (or both where required). Chrome
does not currently enforce this, but will have to in order
to be spec compliant.

BUG=442566
TEST=updated tests pass

Review-Url: https://codereview.chromium.org/2084053002
Cr-Original-Commit-Position: refs/heads/master@{#406687}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51111720ce9dc84fce16319a239126b2c69ca6da