bluetooth: Basic Web Bluetooth Fuzzer
Adds basic structure of the fuzzer, a base template and some tests.

The fuzzer in this file uses two variations of calls to requestDevice to
generate a random sequence of calls. It then replaces instances of
'TRANSFORM_REQUEST_DEVICE_OPTIONS' with object with a single filter with a
single random service.

BUG=570049

Review-Url: https://codereview.chromium.org/2166463002
Cr-Original-Commit-Position: refs/heads/master@{#406940}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 381085b764c743ca32c64168e1f0337050179d03