Simplify OriginTrialContext and the way it validates tokens.
The way OriginTrialContext and the WebTrialTokenValidator API was
designed made a lot of sense when the bindings were still done in such
a way that only when a feature is first used we validate if there are
tokens for that feature. But now that we check every feature in every
context the current design is both not very efficient, and also
needlessly complex.

This simplifies things a lot by validating tokens as soon as AddToken
is called (as all tokens will be validated anyway). This way we only
have to validate each token once rather than once for each feature.

Also gets rid of the old histograms as they weren't very meaningful
anymore: The FeatureEnabled was logged regardless of if a feature was
actually being used on a page, making the data hard to reason about. And
the MessageGenerated histogram was 100% NotRequested after the bindings
changes, making it not useful at all.

BUG=627942

Review-Url: https://codereview.chromium.org/2123323004
Cr-Original-Commit-Position: refs/heads/master@{#407056}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d1c0acb152de57e2d1db5ca9febd3aa67664c90