FileAPI: Remove TODOs around readAsBinaryString - it's back in the spec
FileReader.readAsBinaryString() was removed from the spec after typed
arrays became a thing. But all browsers kept shipping it, and it looks
risky to remove, so it was added back to the spec. Remove the use
counters and TODOs in the IDL files.

R=foolip@chromium.org

Review-Url: https://codereview.chromium.org/2171953002
Cr-Original-Commit-Position: refs/heads/master@{#407202}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f522563b1a430b66d0e1c61cc9629c04d7550c02