[css-flexbox] Some more code refactoring
Add a constructFlexItem function that does the computations
for creating a FlexItem, and returns it. Also add a couple more
cached pieces of data to FlexItem which simplifies code in
a couple of places.

constructFlexItem will also be useful for implementing the
new intrinsic size algorithm.

R=eae@chromium.org,dgrogan@chromium.org
TEST=n/a, no change in behavior

Review-Url: https://codereview.chromium.org/2170533002
Cr-Original-Commit-Position: refs/heads/master@{#407241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3359cdec92d2eb3b61738313ecdae2d60de2e160