Notify the Blink client synchronously if the initial doc is accessed.
This avoids problems with the timer being delayed by nested message
loops and ScopedPageLoadDeferrers, where code in the nested message
loop might access the document and not generate a notification.

The timer was originally added to prevent re-entrant calls to V8,
but that can now be prevented with ScriptForbiddenScope.

BUG=629542
TEST=See bug for repro steps.

Review-Url: https://codereview.chromium.org/2171063002
Cr-Original-Commit-Position: refs/heads/master@{#407269}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00314989401bfee5ffcb5e579162071c86e72f61