Load HRTF database only when the panner is set to HRTF
Currently, the creation of any panner node causes the HRTF database to
be loaded.  However, the default panning model is "equalpower" (about
99.8%) so loading the database is a waste of time and memory.
(According to https://codereview.chromium.org/176683003/, it takes
30MB to hold the database.)

Thus, only load the HRTF database when the panning model is set to
HRTF.

Can't test this easily, but did verify with a debugger that the
database is only created when the panning model is set to HRTF and not
before.

BUG=592787
TEST=none

Review-Url: https://codereview.chromium.org/2171703003
Cr-Original-Commit-Position: refs/heads/master@{#407307}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6610df622ffb33ade51e5ccdb92b7d6b0e2b81ce