Fix networking_private IDL
Minor fixes to the networkingPrivate IDL.
This produces a non-fatal but confusing error in the
Settings UI.

ManagedDOMString[] is invalid, ManagedDOMStringList
should be used instead. (The managed property should
be an objetc with each entry, e.g. 'Active', 'DevicePolicy',
etc, as a separate array of strings, instead of an array of
objects).

BUG=none

Review-Url: https://codereview.chromium.org/2182633002
Cr-Original-Commit-Position: refs/heads/master@{#407583}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 984a730360486f3a02130e81d1ebe6daf036bf52