Add layout test to make sure origin clean flag is preserved
This CL adds a layout test to make sure that the ImageBitmap's origin
clean flag is preserved after ImageBitmap is being transfered or
structured-cloned.

The test works like this: load a cross-origin image, and create an
ImageBitmap from that image. The promise of creating the ImageBitmap
should be resolved but the ImageBitmap is tainted. We then transfer the
ImageBitmap to the worker, and worker structure-clones the ImageBitmap
to the main. We then exam to make sure that the ImageBitmap is still
tained.

BUG=623406

Review-Url: https://codereview.chromium.org/2178513002
Cr-Original-Commit-Position: refs/heads/master@{#407831}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a8242617ffe25ee8738151c768ee41af29c3695