Toggle GPU raster heuristics based on viewportEnabled.
On Android, viewportEnabled can be true and viewportMetaEnabled false
when "request desktop site" is enabled.  We still don't want to turn on
GPU raster unconditionally in this case.

The previous implementation checking viewportMetaEnabled worked largely
by coincidence due to lucky fallthrough and double-setting; change to
viewportEnabled instead which is better aligned with the intent, and add
a unit test for this case.

BUG=

Review-Url: https://codereview.chromium.org/2127023002
Cr-Original-Commit-Position: refs/heads/master@{#407997}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a7a5ed3257d5b8ed2d3f18fb090f5e5bf3b80aa